Button name changed to home-extent-button
