+++ demo/src/app/geo/hover/hover.component.html
@@ -15,7 +15,7 @@
   [igoHoverFeatureEnabled]="true">
     <igo-zoom-button [map]="map" color="primary"></igo-zoom-button>
     <igo-geolocate-button [map]="map" color="primary"></igo-geolocate-button>
-    <igo-initial-button [map]="map" color="primary"></igo-initial-button>
+    <igo-home-extent-button [map]="map" color="primary"></igo-home-extent-button>
     <igo-rotation-button [map]="map" [showIfNoRotation]="true" color="primary"></igo-rotation-button>
   </igo-map-browser>
