+++ demo/src/app/geo/simple-map/simple-map.component.html
@@ -16,7 +16,7 @@
     <igo-info-section [infoContent]="pointerCoord"></igo-info-section>
     <igo-zoom-button [map]="map" color="primary"></igo-zoom-button>
     <igo-geolocate-button [map]="map" color="primary"></igo-geolocate-button>
-    <igo-initial-button [map]="map" color="primary"></igo-initial-button>
+    <igo-home-extent-button [map]="map" color="primary"></igo-home-extent-button>
     <igo-rotation-button [map]="map" [showIfNoRotation]="true" color="primary"></igo-rotation-button>
   </igo-map-browser>
