+++ packages/geo/src/lib/map/home-extent-button/home-extent-button.component.html
@@ -1,6 +1,6 @@
-<div class="igo-initial-button-container"><button
+<div class="igo-home-extent-button-container"><button
       mat-icon-button
-      [matTooltip]="'igo.geo.mapButtons.initial' | translate"
+      [matTooltip]="'igo.geo.mapButtons.home-extent' | translate"
             matTooltipPosition="left"
       [color]="color"
       (click)="onToggleClick($event)"><mat-icon svgIcon = "home-map-marker"></mat-icon></button></div>
