+++ packages/geo/src/lib/map/home-extent-button/home-extent-button.component.scss
@@ -1,6 +1,6 @@
 @import '../../../../../core/src/style/partial/core.variables';
 
-.igo-initial-button-container {
+.igo-home-extent-button-container {
   width: $igo-icon-size;
   background-color: #fff;
   &:hover {
