+++ packages/geo/src/lib/map/home-extent-button/index.ts
@@ -1,0 +1,1 @@
+export * from './home-extent-button.component';
