+++ packages/geo/src/lib/map/index.ts
@@ -4,7 +4,7 @@ export * from './map-browser';
 export * from './zoom-button';
 export * from './menu-button';
 export * from './geolocate-button';
-export * from './initial-button';
+export * from './home-extent-button';
 export * from './offline-button';
 export * from './baselayers-switcher';
 export * from './rotation-button';
