--- packages/geo/src/lib/map/initial-button/index.ts
@@ -1,1 +1,0 @@
-export * from './initial-button.component';
