+++ packages/geo/src/lib/map/map.module.ts
@@ -8,7 +8,7 @@ import { IgoConfirmDialogModule } from '@igo2/common';
 import { MapBrowserComponent } from './map-browser/map-browser.component';
 import { ZoomButtonComponent } from './zoom-button/zoom-button.component';
 import { GeolocateButtonComponent } from './geolocate-button/geolocate-button.component';
-import { InitialButtonComponent } from './initial-button/initial-button.component';
+import { HomeExtentButtonComponent } from './home-extent-button/home-extent-button.component';
 import { RotationButtonComponent } from './rotation-button/rotation-button.component';
 import { BaseLayersSwitcherComponent } from './baselayers-switcher/baselayers-switcher.component';
 import { MiniBaseMapComponent } from './baselayers-switcher/mini-basemap.component';
@@ -34,7 +34,7 @@ import { InfoSectionComponent } from './info-section/info-section.component';
     MapBrowserComponent,
     ZoomButtonComponent,
     GeolocateButtonComponent,
-    InitialButtonComponent,
+    HomeExtentButtonComponent,
     RotationButtonComponent,
     InfoSectionComponent,
     BaseLayersSwitcherComponent,
@@ -51,7 +51,7 @@ import { InfoSectionComponent } from './info-section/info-section.component';
     MapBrowserComponent,
     ZoomButtonComponent,
     GeolocateButtonComponent,
-    InitialButtonComponent,
+    HomeExtentButtonComponent,
     RotationButtonComponent,
     InfoSectionComponent,
     BaseLayersSwitcherComponent,
