+++ packages/geo/src/locale/en.geo.json
@@ -265,6 +265,7 @@
       "mapButtons": {
         "baselayerSwitcher": "Change base layer",
         "geolocate": "Geolocate",
+        "home-extent": "Go to the home map extent",
         "offline": "Offline mode",
         "online": "Online mode",
         "zoomIn": "Zoom in ({{zoom}})",
